UPDATE: Following a year of experiments and feedback a number of changes are being made to OWIDs prior to deployment in production SWAN implementations.
OWID is now compatible with the Source data entity explained in the OneKey project.

Signer is now used in place of the word Creator to indicate the entity which captured or generated the data and OWID.

The payload was originally included in the OWID data structure. This has now been removed as it reduces flexibility. The change does mean implementors will need to be more careful about how they extract the data that forms the OWID and explain this to receivers. This change reduces the binary format representation of the OWID to a minimum of 73 bytes. The term associated data is now used when describing the data that the OWID relates to.

Removed the restriction to store OWIDs as binary data structures. JSON and other formats are now available.

Made the return of the terms and conditions URL in the signer response a mandatory field to ensure that receivers understand the legal basis on which the associated data was captured or generated. OWID hosts will need to use multiple domains to host multiple terms and conditions.